refactor: use common types on ProcessMembership
