+++ ProcessMembership/handler.ts
@@ -6,27 +6,10 @@ import { readableReport } from "@pagopa/ts-commons/lib/reporters";
 import { NotImplementedError, ValidationError } from "../models/error";
 import { withJsonInput } from "../utils/misc";
 import { Dao } from "../models/dao";
-
-type IpaCode = t.TypeOf<typeof IpaCode>;
-const IpaCode = t.string;
+import { IContract, IpaCode } from "../models/types";
 
 type QueueItem = t.TypeOf<typeof QueueItem>;
 const QueueItem = t.type({
-  ipaCode: IpaCode
-});
-
-type Delegate = t.TypeOf<typeof Contract>;
-const Delegate = t.type({
-  email: t.string,
-  fiscalCode: t.string,
-  id: t.string,
-  role: t.string
-});
-
-type Contract = t.TypeOf<typeof Contract>;
-const Contract = t.type({
-  delegates: t.readonlyArray(Delegate),
-  id: t.string,
   ipaCode: IpaCode
 });
 
@@ -38,22 +21,22 @@ type SelfCareMembershipClaimParams = {
 // retrieve the list of contracts relative to a Membership, identified by its ipa code
 const fetchContractsByIpaCode = (_dao: Dao) => (
   _ipaCode: IpaCode
-): TE.TaskEither<Error, ReadonlyArray<Contract>> =>
+): TE.TaskEither<Error, ReadonlyArray<IContract>> =>
   TE.left(
     new NotImplementedError("fetchContractsByIpaCode() - Not implemented yet")
   );
 
 // Given a list of contracts, get the sublist of the ones we should consider in our process
-const selectContract = (contracts: ReadonlyArray<Contract>): Contract =>
+const selectContract = (contracts: ReadonlyArray<IContract>): IContract =>
   contracts[0];
 
 // Check if a person with manager role has been declared in at least one of the contracts
-const hasManager = (_contract: Contract): boolean => false;
+const hasManager = (_contract: IContract): boolean => false;
 
 // Prepare data to be sent to SelfCare
 const composeSelfCareMembershipClaim = (
   _ipaCode: IpaCode,
-  _contract: Contract
+  _contract: IContract
 ): SelfCareMembershipClaimParams => ({});
 
 // Submit the claim to SelfCare to import the memebership
@@ -69,13 +52,13 @@ const submitMembershipClaimToSelfcare = (
 // Save that the memebeship is not meant to be processed by the current business logic
 const markMembershipAsDiscarded = (_dao: Dao) => (
   _ipaCode: IpaCode
-): TE.TaskEither<Error, ReadonlyArray<Contract>> =>
+): TE.TaskEither<Error, ReadonlyArray<IContract>> =>
   TE.left(new NotImplementedError("discardMembership() - Not implemented yet"));
 
 // Save that the memebeship has been correctly claimed to SelfCare
 const markMembershipAsCompleted = (_dao: Dao) => (
   _ipaCode: IpaCode
-): TE.TaskEither<Error, ReadonlyArray<Contract>> =>
+): TE.TaskEither<Error, ReadonlyArray<IContract>> =>
   TE.left(
     new NotImplementedError("markMembershipAsCompleted() - Not implemented yet")
   );
