+++ models/types.ts
@@ -1,10 +1,13 @@
 import * as t from "io-ts";
 
+export type IpaCode = t.TypeOf<typeof IpaCode>;
+export const IpaCode = t.string;
+
 export type IMembership = t.TypeOf<typeof IMembership>;
 export const IMembership = t.intersection([
   t.type({
     id: t.string,
-    ipaCode: t.string,
+    ipaCode: IpaCode,
     mainInstitution: t.boolean,
     status: t.string
   }),
@@ -24,9 +27,17 @@ export const IContract = t.type({
   attachment: IAttachment,
   emailDate: t.string,
   id: t.string,
-  ipaCode: t.string,
+  ipaCode: IpaCode,
   version: t.string
 });
+
+/* type Delegate = t.TypeOf<typeof Delegate>;
+const Delegate = t.type({
+  email: t.string,
+  fiscalCode: t.string,
+  id: t.string,
+  role: t.string
+}); */
 
 export type TipoContratto = t.TypeOf<typeof TipoContratto>;
 export const TipoContratto = t.union([
